refactor: split src/context.rs into src/context/ module directory
Extract budget helpers (token estimation, section truncation, budget
fitting) into src/context/budget.rs and generic rendering helpers
(section creation, message headers, labels, sanitization) into
src/context/render.rs. The main orchestrators and domain-specific
builders remain in src/context/mod.rs.

This is a mechanical refactor toward the src/context/ module split
requested in #1892. No behavioral changes; all 73 existing context
tests pass plus one new regression test verifying the submodule
cooperation boundary.

Closes #1892
